+++ src/app/utils/filesystem-browser-model.js
@@ -1135,12 +1135,14 @@ export default BaseBrowserModel.extend(DownloadInBrowser, {
       //   width: 160,
       // }),
     });
+    this.set('columnsOrder', ['size', 'modification']);
     if (this.isOwnerVisible) {
       this.set('columns.owner', EmberObject.create({
         isVisible: true,
         isEnabled: true,
         width: 200,
       }));
+      this.columnsOrder.push('owner');
     }
     this._super(...arguments);
   },